% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{census.2000_2010}
\alias{census.2000_2010}
\title{Population data - US - 2000 and 2010}
\format{A data frame with 3143 rows and 5 variables}
\description{
A dataset containing population data of the United States - 2000 and 2010.
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{dataset}

