\docType{data}
\name{corporate.payment}
\alias{corporate.payment}
\title{Corporate payments of a West Coast utility company - 2010}
\format{A data frame with 189470 rows and 4 variables}
\description{
A dataset of the 2010's payments data of a division of a
West Coast utility company.
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for
Forensic Accounting, Auditing and Fraud Detection. Wiley
and Sons: New Jersey.
}
\keyword{dataset}

