% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{fibonacci.500}
\alias{fibonacci.500}
\title{The first 500 Fibonacci numbers.}
\format{A data frame with 500 rows and 1 variable.}
\description{
A dataset containing the first 500 Fibonacci numbers.
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{dataset}
