% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{gm.payments}
\alias{gm.payments}
\title{General Motors payments.}
\format{A data frame with 15300 rows and 1 variable.}
\description{
A dataset containing the amounts paid to vendors for the 90 days preceding liquidation - 2009.
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey, pp. 253-255.
}
\keyword{dataset}
