% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{invoices.paid}
\alias{invoices.paid}
\title{Invoices paid by a utility company.}
\format{A data frame with 189470 rows and 5 variables.}
\description{
A dataset containing the invoices paid by a utility company.
}
\references{
Nigrini, M. J. (2011). Forensic Analytics: Methods and Techniques for Forensic Accounting Investigations. Wiley and Sons: New Jersey.
}
\keyword{dataset}
