% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{journal.entry}
\alias{journal.entry}
\title{Journal entries in a company's accounting system.}
\format{A data frame with 154935 rows and 1 variable.}
\description{
A dataset related to the 2005 journal entries in a company's accounting system.
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{dataset}
