% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.functions.R
\name{ks}
\alias{ks}
\title{Gets the Kolmogorov-Smirnov test of a Benford object}
\usage{
ks(bfd)
}
\arguments{
\item{bfd}{an object of class "Benford". See \code{\link{benford}}.}
}
\value{
A list with class "htest" containing the results of the Kolmogorov-Smirnov test.
}
\description{
It gets the Kolmogorov-Smirnov test for a Benford object.
See the section value of \code{\link{benford}}.
}
\examples{
data(census.2009) #gets data
c2009 <- benford(census.2009$pop.2009) #generates benford object
ks(c2009) # equivalent to c2009$stats$ks.test
}
