% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{lakes.perimeter}
\alias{lakes.perimeter}
\title{Perimeter of lakes arround the world}
\format{A data frame with 248607 rows and 1 variable.}
\description{
A dataset of the perimeter of the lakes arround the water from
the global lakes and wetlands database (GLWD).
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{dataset}

