% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{last.two.digits}
\alias{last.two.digits}
\title{Extracts the last two digits from the data}
\usage{
last.two.digits(data, sign = "positive")
}
\arguments{
\item{data}{a numeric vector.}

\item{sign}{The default value for sign is "positive" and it analyzes only data greater than zero. 
There are also the options "negative" and "both" that will analyze only negative values or both positive and negative values of the data,
respectively. For large datasets with both positive and negative numbers, 
it is usually recommended to perform a separate analysis for each group,
for the incentives to manipulate the numbers are usually different.}
}
\value{
A data.frame with the data and the last digits.
}
\description{
It extracts the last two digits from the data.

This function is used by the main function of the package \code{\link{benford}} to extract the 
ast two digits of the data.
}
