% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions-new-code-2.R
\name{MAD.conformity}
\alias{MAD.conformity}
\title{MAD conformity to Benford's Law using the MAD}
\usage{
MAD.conformity(MAD = NULL, digits.used = c("First Digit",
  "Second Digit", "First-Two Digits", "First-Three Digits"))
}
\arguments{
\item{MAD}{The mean absolute deviation, as computed in the function \code{\link{benford}}}

\item{digits.used}{How many digits used in the analysis.}
}
\value{
A list with the MAD, digits.used and the conformity level.
}
\description{
This function checks the MAD agains the confomity criteria proposed by Nigrini (2012).
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
