% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{madoff.returns}
\alias{madoff.returns}
\title{The potential amount of the exposure of a partial list released by Wall Street Journal (WSJ) of the Madoff victims - 2009.}
\format{A data frame with 215 rows and 1 variable.}
\description{
A dataset containing the potential amount of the exposure of a partial list of the Madoff victims.
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{dataset}
