% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plot.Benford}
\alias{plot.Benford}
\title{Plot method for Benford Analysis}
\usage{
\method{plot}{Benford}(x, select = c("digits", "obs vs exp", "summation",
  "diff vs ex summation", "last two digits", "second order", "mantissa",
  "mantissa arc"), except = NULL, multiple = TRUE,
  col.bar = "lightblue", err.bounds = FALSE, alpha = 0.05,
  grid = TRUE, mfrow = NULL, exp.benford = TRUE, freq = TRUE, ...)
}
\arguments{
\item{x}{a "Benford" object}

\item{select}{it specifies the order and which plots are going to be plotted. If NULL, the parameter except is used.
Currently, you can choose from 13 plots:\itemize{
\item \code{"digits": }{a barplot concerning the observation frequencies with respect to its first-order digits, comparing it with the value expected by the Benford's Law. }
\item \code{"rootogram digits": }{the hanging rootogram is variation of the histogram with the vertical axis showing the frequencies hanging from fitted or expected frequencies (here by Benford's Law) so that the discrepancies are visualized against a straight line (the axis) rather than against a curve.}
\item \code{"obs vs exp": }{a scatter plot of the observed frequencies versus expected frequencies.}
\item \code{"second order": }{a barplot concerning the count for the ordered data difference}
\item \code{"rootogram second order": }{is similar to \code{"rootogram digits"}. The observed frequencies of the ordered data difference are displayed as bars and the expected frequencies by the Benford's Law as a line.}
\item \code{"last two digits": }{a barplot concerning the observation frequencies with respect to its last two digits, comparing it with the value expected by the Benford's Law. }
\item \code{"summation": }{a barplot concerning the deviations of summation values from expected values.}
\item \code{"mantissa": }{a plot concerning the ordered mantissas.}
\item \code{"mantissa arc": }{a plot of mantissa arc test.}
\item \code{"chi square": }{a needle plot concerns the chi-squared statistics consisting of deviations of observed values from expected values.}
\item \code{"abs diff": }{a needle plot concerning the absolute difference of observed values from expected values.}
\item \code{"ex summation": }{a needle plot concerning the excess summation of observed values from expected values.}
\item \code{"diff vs ex summation": }{a scatter plot of the difference between observed and observed frequencies versus excess summations.}
}}

\item{except}{it specifies which plots are not going to be plotted. If NULL, the parameter select is used.}

\item{multiple}{logical; if TRUE, all plots are grouped in the same window.}

\item{col.bar}{a color to be used to fill the bars. The default is lightblue.}

\item{err.bounds}{logical; if TRUE, the upper and lower error bounds are draw. The error bounds indicate the binomial root mean square error.}

\item{alpha}{it specifies level of confidence interval. The defaults to 95 percent confidence interval,i.e., the error bounds will represent 1.96 standard error from the expected count by Benford's Law.}

\item{grid}{logical; if TRUE, adds an rectangular grid to plot.}

\item{mfrow}{A vector of the form c(`nr`, `nc`). Subsequent figures will be drawn in an `nr`-by-`nc` array on the device by columns (`mfcol`), or rows (`mfrow`), respectively. For more details see `?par()`.}

\item{exp.benford}{logical; if TRUE, adds a line representing the expected frequencies by Berford's Law.}

\item{freq}{logical; if TRUE, the plot is a representation of counts; if FALSE, relative frequencies are plotted (so that the plot has a total area of one). Defaults to `TRUE`.}

\item{...}{arguments to be passed to generic plot functions,}
}
\value{
Plots the Benford object.
}
\description{
The \code{plot} method for "Benford" objects.
}
\details{
If both \code{select} and \code{except} arguments have been provided, but only \code{select} will be considered and \code{except} ignored.
For more flexibility in plotting, use (`separate plots here`)...
}
