\name{print.Benford}
\alias{print.Benford}
\title{Print method for Benford Analysis}
\usage{
\method{print}{Benford}(x, how.many=5, ...)
}
\arguments{
  \item{x}{a "Benford" object.}

  \item{how.many}{a number that defines how many of the
  biggest absolute differences to show.}

  \item{...}{arguments to be passed to generic print
  functions.}
}
\value{
Prints the Benford object.
}
\description{
The \code{print} method for "Benford" objects.
}

