% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{purchasing.cards.2010}
\alias{purchasing.cards.2010}
\title{Purchasing cards transactions - 2010.}
\format{A data frame with 95076 rows and 8 variables.}
\description{
A dataset containing the card transactions for a government entity - 2010.
}
\references{
Nigrini, M. J. (2011). Forensic Analytics: Methods and Techniques for Forensic Accounting Investigations. Wiley and Sons: New Jersey.
}
\keyword{dataset}
