% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.functions.R
\name{summary.Benford}
\alias{summary.Benford}
\title{Summary method for Benford Analysis}
\usage{
\method{summary}{Benford}(object, what = c("first digits", "summation",
  "last two digits", "second order"), sort.by = "abs diff",
  freq = TRUE, ...)
}
\arguments{
\item{object}{a "Benford" object}

\item{what}{options: "first digits", "summation", "last two digits", "second order".}

\item{sort.by}{(not implemented)}

\item{freq}{(not implemented)}

\item{...}{arguments to be passed to generic summary function.}
}
\value{
Summary of the Benford object.
}
\description{
The \code{summary} method for "Benford" objects.
}
