% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{taxable.incomes.1978}
\alias{taxable.incomes.1978}
\title{Taxable Income 1978}
\format{A data frame with 157518 rows and 1 variable.}
\description{
Taxable Incomes of the 1978 Individual Tax Model File (ITMF).
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{dataset}

