% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{truncDec}
\alias{truncDec}
\title{Trunc decimal values.}
\usage{
truncDec(x, n.dec = 2)
}
\arguments{
\item{x}{a numeric vector.}

\item{n.dec}{integer indicating the number of decimal places (trunc) to be used.}
}
\value{
A numeric vector with the data with deciml values truncated.
}
\description{
It trunc decimal values from the data.

This function is used by the main function of the package \code{\link{benford}}
to trunc the values in its ndec argument to the specified number of decimal places.
}
