% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plot.Benford}
\alias{plot.Benford}
\title{Plot method for Benford Analysis}
\usage{
\method{plot}{Benford}(x, select = c("digits", "second order",
  "summation", "chi squared", "ex summation"), except = NULL,
  multiple = TRUE, col.bar = "turquoise3", err.bounds = FALSE,
  alpha = 0.05, grid = TRUE, mfrow = NULL, exp.benford = TRUE,
  freq = TRUE, ...)
}
\arguments{
\item{x}{a "Benford" object}

\item{select}{it specifies the order and which plots are going to be plotted. If NULL, the parameter except is used.}

\item{except}{it specifies which plots are not going to be plotted. If NULL, the parameter select is used.
Currently, you can choose from 10 plots: "digits", "rootogram digits", "second order", "rootogram second order", "last two digits", "summation",
"mantissa", "chi square", "abs diff", "ex summation". If you want to plot all, just
put except = "none". The default is not to plot the "mantissa" and "abs diff". If you want to plot all, just
put except = "all"}

\item{multiple}{logical; if TRUE, all plots are grouped in the same window.}

\item{col.bar}{a color to be used to fill the bars. The default is lightblue.}

\item{err.bounds}{logical; if TRUE, the upper and lower error bounds are draw. The error bounds indicate the binomial root mean square error.}

\item{alpha}{it specifies level of confidence interval. The defaults to 95 percent confidence interval,i.e., the error bounds will represent 1.96 standard error from the expected count by Benford's Law.}

\item{grid}{logical; if TRUE, adds an rectangular grid to plot.}

\item{mfrow}{A vector of the form c(`nr`, `nc`). Subsequent figures will be drawn in an `nr`-by-`nc` array on the device by columns (`mfcol`), or rows (`mfrow`), respectively. For more details see `?par()`.}

\item{exp.benford}{logical; if TRUE, adds a line representing the expected frequencies by Berford's Law.}

\item{freq}{logical; if TRUE, the plot is a representation of counts; if FALSE, relative frequencies are plotted (so that the plot has a total area of one). Defaults to `TRUE`.}

\item{...}{arguments to be passed to generic plot functions,}
}
\value{
Plots the Benford object.
}
\description{
The \code{plot} method for "Benford" objects.
}
\details{
If both \code{select} and \code{except} arguments have been provided, but only \code{select} will be considered and \code{except} ignored.
}
