% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{sino.forest}
\alias{sino.forest}
\title{Financial Statements of Sino Forest Corporation's 2010 Report}
\format{A data frame with 772 rows and 1 variable.}
\description{
Financial Statements numbers of Sino Forest Corporation's 2010 Report.
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{dataset}
